+++ packages/ai/errors/index.ts
@@ -14,6 +14,7 @@ export {
 
 export { InvalidArgumentError } from './invalid-argument-error';
 export { InvalidToolArgumentsError } from './invalid-tool-arguments-error';
+export { NoImageGeneratedError } from './no-image-generated-error';
 export { NoObjectGeneratedError } from './no-object-generated-error';
 export { NoOutputSpecifiedError } from './no-output-specified-error';
 export { NoSuchToolError } from './no-such-tool-error';