+++ CHANGELOG.md
@@ -20,6 +20,7 @@ You can find the changelogs for the individual packages in their respective `CHA
 - [@ai-sdk/mistral](./packages/mistral/CHANGELOG.md)
 - [@ai-sdk/openai](./packages/openai/CHANGELOG.md)
 - [@ai-sdk/openai-compatible](./packages/openai-compatible/CHANGELOG.md)
+- [@ai-sdk/perplexity](./packages/perplexity/CHANGELOG.md)
 - [@ai-sdk/togetherai](./packages/togetherai/CHANGELOG.md)
 - [@ai-sdk/xai](./packages/xai/CHANGELOG.md)