+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,37 @@
 
 ### Patch Changes
 
+- Updated dependencies [9ce598c]
+  - ai@4.1.5
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [caaad11]
+- Updated dependencies [caaad11]
+  - ai@4.1.4
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [7f30a77]
+- Updated dependencies [4298996]
+  - ai@4.1.3
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [3c5fafa]
+- Updated dependencies [3a58a2e]
+  - ai@4.1.2
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [0a699f1]
   - ai@4.1.1