+++ packages/cerebras/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @ai-sdk/cerebras
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [361fd08]
+  - @ai-sdk/openai-compatible@0.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [ed012d2]
+- Updated dependencies [3a58a2e]
+  - @ai-sdk/openai-compatible@0.1.2
+  - @ai-sdk/provider-utils@2.1.2
+  - @ai-sdk/provider@1.0.6
+
 ## 0.1.1
 
 ### Patch Changes