+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/google-vertex
 
+## 2.1.2
+
+### Patch Changes
+
+- 3a58a2e: feat (ai/core): throw NoImageGeneratedError from generateImage when no predictions are returned.
+- Updated dependencies [ed012d2]
+- Updated dependencies [3a58a2e]
+  - @ai-sdk/provider-utils@2.1.2
+  - @ai-sdk/provider@1.0.6
+  - @ai-sdk/anthropic@1.1.2
+  - @ai-sdk/google@1.1.2
+
 ## 2.1.1
 
 ### Patch Changes