+++ packages/swarm/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @ai-sdk/swarm
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [9ce598c]
+  - ai@4.1.5
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [caaad11]
+- Updated dependencies [caaad11]
+  - ai@4.1.4
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [7f30a77]
+- Updated dependencies [4298996]
+  - ai@4.1.3
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [3c5fafa]
+- Updated dependencies [3a58a2e]
+  - ai@4.1.2
+
 ## 0.1.1
 
 ### Patch Changes