+++ packages/ui-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/ui-utils
 
+## 1.1.3
+
+### Patch Changes
+
+- 9ce598c: feat (ai/ui): add reasoning support to useChat
+
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [ed012d2]
+- Updated dependencies [3a58a2e]
+  - @ai-sdk/provider-utils@2.1.2
+  - @ai-sdk/provider@1.0.6
+
 ## 1.1.1
 
 ### Patch Changes